Rename 'r#async' module to 'not_sync'.
Best idea I've got!
