+++ src/call/client.rs
@@ -19,7 +19,7 @@ use futures::{Async, AsyncSink, Future, Poll, Sink, StartSend, Stream};
 use crate::grpc_sys;
 
 use super::{ShareCall, ShareCallHolder, SinkBase, WriteFlags};
-use crate::r#async::{BatchFuture, BatchType, SpinLock};
+use crate::not_sync::{BatchFuture, BatchType, SpinLock};
 use crate::call::{check_run, Call, MessageReader, Method};
 use crate::channel::Channel;
 use crate::codec::{DeserializeFn, SerializeFn};
